requirements.txt
appdirs==1.4.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Cython==0.29.15
distlib==0.3.0
Django==2.2
django-autocomplete-light==3.5.1
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-mdeditor==0.1.17
django-redis==4.11.0
django-silk==4.0.0
django-simpleui==3.9
djangorestframework==3.11.0
djdt-flamegraph==0.2.13
filelock==3.0.12
get==2019.4.13
gprof2dot==2019.11.30
hiredis==1.0.1
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.1.0
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==1.4.6
numpy==1.18.1
panda==0.3.1
pandas==1.0.1
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
pycodestyle==2.5.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
query-string==2019.4.13
redis==3.4.1
request==2019.4.13
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.7
zipp==3.0.0
